Extract pdf rendering function into a helper module
